+++ CVGS/Controllers/ReviewsController.cs
@@ -17,16 +17,18 @@ namespace CVGS.Controllers
         // GET: Reviews/Create
         public ActionResult Create(int? gameId)
         {
+            // Redirect unauthenticated members
+            var memberId = this.Session["MemberId"];
+            if (memberId == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
             if (gameId == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            if (Session["MemberId"] == null)
-            {
-                return RedirectToAction("Index", "Login");
-            }
-            int memberId = (int)Session["MemberId"];
-
+            
             if (db.REVIEWs.Find(memberId, gameId) != null)
             {
                 // Member has already reviewed this game; redirect to edit
@@ -37,7 +39,7 @@ namespace CVGS.Controllers
             {
                 GameId = (int)gameId,
                 GAME = db.GAMEs.Find(gameId),
-                MemberId = (int)Session["MemberId"]
+                MemberId = (int)memberId
             };
 
             return View(review);
@@ -50,6 +52,13 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "MemberId,GameId,ReviewText,Rating")] REVIEW review)
         {
+            // Redirect unauthenticated members
+            var memberId = this.Session["MemberId"];
+            if (memberId == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
             if (ModelState.IsValid)
             {
                 review.DateCreated = DateTime.Now;
@@ -64,18 +73,22 @@ namespace CVGS.Controllers
         // GET: Reviews/Edit/5
         public ActionResult Edit(int? memberId, int? gameId)
         {
+            // Redirect unauthenticated members
+            if (this.Session["MemberId"] == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
             if (memberId == null || gameId == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            if (Session["MemberId"] == null)
-            {
-                return RedirectToAction("Index", "Login");
-            }
+
             if (memberId != (int)Session["MemberId"])
             {
                 return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             }
+
             REVIEW review = db.REVIEWs.Find(memberId, gameId);
             if (review == null)
             {
@@ -91,6 +104,13 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "MemberId,GameId,ReviewText,Rating,DateCreated")] REVIEW review)
         {
+            // Redirect unauthenticated members
+            var memberId = this.Session["MemberId"];
+            if (memberId == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
             if (ModelState.IsValid)
             {
                 review.DateModified = DateTime.Now;
@@ -104,14 +124,22 @@ namespace CVGS.Controllers
         // GET: Reviews/Delete/5
         public ActionResult Delete(int? memberId, int? gameId)
         {
+            // Redirect unauthenticated members
+            if (this.Session["MemberId"] == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
             if (memberId == null || gameId == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             if (memberId != (int)Session["MemberId"] && (string)Session["MemberRole"] != "Employee" && (string)Session["MemberRole"] != "Admin" )
             {
                 return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             }
+
             REVIEW review = db.REVIEWs.Find(memberId, gameId);
             if (review == null)
             {
@@ -125,7 +153,14 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed([Bind(Include = "MemberId,GameId")] REVIEW reviewToDelete)
         {
-            //TODO Implement Delete, probably with a view model
+            // Redirect unauthenticated members
+            var memberId = this.Session["MemberId"];
+            if (memberId == null)
+            {
+                return RedirectToAction("Index", "Login"); ;
+            }
+
+            //TODO: Implement Delete, probably with a view model
             REVIEW review = db.REVIEWs.Find(reviewToDelete.MemberId, reviewToDelete.GameId);
             db.REVIEWs.Remove(review);
             db.SaveChanges();
